Update GoogleSpeechProcessor.java
Getting errors when multiple result segments come back from Google (session.transfer is called more than once). I've updated the processor to handle those but adding the metrics to the FlowFile doesn't really make sense in that context. We will have to compute aggregate metrics if we want to add them to the FlowFile.

Additionally, I found the S2T works with 8000Hz sample rate. We should probably put this into the processor as a configuration setting, I've hard-coded for now to get it working.